Final Commit added 2 spaces to invoke CICD

+++ stop_container.sh
@@ -2,5 +2,9 @@
 set -e
 
 # Stop the running container (if any)
-containerid=$(docker ps | awk -F " " '{print $1}')
-docker rm -f $containerid
+containerid=$(docker ps | awk 'NR==2{print $1}')
+if [ ! -z "$containerid" ]; then
+    docker rm -f $containerid
+else
+    echo "No running containers found."
+fi
